[IMP] shipment_advice: test with normal stock user
When testing with `shopfloor_delivery_shipment` module, there was error
in this module as below. That is because normal stock user is used in
testing. So this commit is to add test with normal stock user to make
sure UI works as expected.

```
odoo.exceptions.AccessError: You are not allowed to access 'Action Window' (ir.actions.act_window) records.

This operation is allowed for the following groups:
        - Administration/Settings

Contact your administrator to request access if necessary.
```